Add a simple optimizer that preevaluates subterms
